refactor: split ui dialog modules
